requirements.txt
twiggy
python-magic
olefile
git+https://github.com/Rafiot/officedissector.git
git+https://github.com/Rafiot/pdfid.git
